[Update cali.py]: delete confirm function
The confirm function was used to keep Windows cmd open after
program finishes or after an error accured. But it's superfluous
and not elegant, so it has been expired for a long time.

In the future maybe I will add GUI for Windows users, so that
they do not need to use this script at command line every time.